requirements.txt
asgiref==3.9.1
dj-database-url==3.0.1
Django==5.2.4
django-ckeditor==6.7.3
django-js-asset==3.1.2
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.1
mysqlclient==2.2.7
PyJWT==2.10.1
PyMySQL==1.1.1
python-dotenv==1.1.1
sqlparse==0.5.3
tzdata==2025.2
